Las to multipoint shapefile

Converts one or more LAS files into MultipointZ vector Shapefiles. When the input parameter is not specified, the tool grids all LAS files contained within the working directory.

Usage

wbt_las_to_multipoint_shapefile(
  input,
  wd = NULL,
  verbose_mode = NULL,
  compress_rasters = NULL,
  command_only = FALSE
)

Arguments

input

Input LiDAR file.

wd

Changes the working directory. Default: NULL will use the value in WhiteboxTools settings, see wbt_wd() for details.

verbose_mode

Sets verbose mode. If verbose mode is FALSE, tools will not print output messages. Default: NULL will use the value in WhiteboxTools settings, see wbt_verbose() for details.

compress_rasters

Sets the flag used by 'WhiteboxTools' to determine whether to use compression for output rasters. Default: NULL will use the value in WhiteboxTools settings, see wbt_compress_rasters() for details.

command_only

Return command that would be executed by system() rather than running tool. Default: FALSE.

Value

Returns the tool text outputs.