Lidar idw interpolation

Interpolates LAS files using an inverse-distance weighted (IDW) scheme. When the input/output parameters are not specified, the tool interpolates all LAS files contained within the working directory.

wbt_lidar_idw_interpolation(
  input,
  output = NULL,
  parameter = "elevation",
  returns = "all",
  resolution = 1,
  weight = 1,
  radius = 2.5,
  exclude_cls = NULL,
  minz = NULL,
  maxz = NULL,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE
)

Arguments

input
Input LiDAR file (including extension).

output
Output raster file (including extension).

parameter
Interpolation parameter; options are 'elevation' (default), 'intensity', 'class', 'return_number', 'number_of_returns', 'scan angle', 'rgb', 'user data'.

returns
Point return types to include; options are 'all' (default), 'last', 'first'.

resolution
Output raster's grid resolution.

weight
IDW weight value.

radius
Search Radius.

exclude_cls
Optional exclude classes from interpolation; Valid class values range from 0 to 18, based on LAS specifications. Example, --exclude_cls='3,4,5,6,7,18'.

minz
Optional minimum elevation for inclusion in interpolation.

maxz
Optional maximum elevation for inclusion in interpolation.

wd
Changes the working directory.

verbose_mode
Sets verbose mode. If verbose mode is False, tools will not print output messages.

compress_rasters
Sets the flag used by WhiteboxTools to determine whether to use compression for output rasters.

Value

Returns the tool text outputs.