Lidar ransac planes

Performs a RANSAC analysis to identify points within a LiDAR point cloud that belong to linear planes.

wbt_lidar_ransac_planes(
  input,
  output,
  radius = 2,
  num_iter = 50,
  num_samples = 5,
  threshold = 0.35,
  model_size = 8,
  max_slope = 80,
  classify = FALSE,
  last_returns = FALSE,
  wd = NULL,
  verbose_mode = FALSE,
  compress_rasters = FALSE
)

Arguments

input
Input LiDAR file.

output
Output LiDAR file.

radius
Search Radius.

num_iter
Number of iterations.

num_samples
Number of sample points on which to build the model.

threshold
Threshold used to determine inlier points.

model_size
Acceptable model size.

max_slope
Maximum planar slope.

classify
Classify points as ground (2) or off-ground (1).

last_returns
Only include last- and only-return points.

wd
Changes the working directory.

verbose_mode
Sets verbose mode. If verbose mode is False, tools will not print output messages.

compress_rasters
Sets the flag used by WhiteboxTools to determine whether to use compression for output rasters.

Value

Returns the tool text outputs.