All available tools in WhiteboxTools.

All available tools in WhiteboxTools.

wbt_list_tools(keywords = NULL)

Arguments

keywords
Keywords may be used to search available tools.

Value

Return all available tools in WhiteboxTools that contain the keywords.

Examples

wbt_list_tools("lidar")

#>  [1] "All 38 Tools containing keywords:"                                                                                                                                                                                             
#>  [2] "ClassifyOverlapPoints: Classifies or filters LAS points in regions of overlapping flight lines."                                                                                                                               
#>  [3] "ClipLidarToPolygon: Clips a LiDAR point cloud to a vector polygon or polygons."                                                                                                                                                
#>  [4] "ErasePolygonFromLidar: Erases (cuts out) a vector polygon or polygons from a LiDAR point cloud."                                                                                                                               
#>  [5] "FilterLidarScanAngles: Removes points in a LAS file with scan angles greater than a threshold."                                                                                                                                
#>  [6] "FindFlightlineEdgePoints: Identifies points along a flightline's edge in a LAS file."                                                                                                                                          
#>  [7] "FlightlineOverlap: Reads a LiDAR (LAS) point file and outputs a raster containing the number of overlapping flight lines in each grid cell."                                                                                   
#>  [8] "LasToAscii: Converts one or more LAS files into ASCII text files."                                                                                                                                                             
#>  [9] "LasToMultipointShapefile: Converts one or more LAS files into MultipointZ vector Shapefiles. When the input parameter is not specified, the tool grids all LAS files contained within the working directory."                  
#> [10] "LasToShapefile: Converts one or more LAS files into a vector Shapefile of POINT ShapeType."                                                                                                                                    
#> [11] "LidarBlockMaximum: Creates a block-maximum raster from an input LAS file. When the input/output parameters are not specified, the tool grids all LAS files contained within the working directory."                            
#> [12] "LidarBlockMinimum: Creates a block-minimum raster from an input LAS file. When the input/output parameters are not specified, the tool grids all LAS files contained within the working directory."                            
#> [13] "LidarClassifySubset: Classifies the values in one LiDAR point cloud that correpond with points in a subset cloud."                                                                                                             
#> [14] "LidarColourize: Adds the red-green-blue colour fields of a LiDAR (LAS) file based on an input image."                                                                                                                          
#> [15] "LidarConstructVectorTIN: Creates a vector triangular irregular network (TIN) fitted to LiDAR points."                                                                                                                          
#> [16] "LidarElevationSlice: Outputs all of the points within a LiDAR (LAS) point file that lie between a specified elevation range."                                                                                                  
#> [17] "LidarGroundPointFilter: Identifies ground points within LiDAR dataset using a slope-based method."                                                                                                                             
#> [18] "LidarHexBinning: Hex-bins a set of LiDAR points."                                                                                                                                                                              
#> [19] "LidarHillshade: Calculates a hillshade value for points within a LAS file and stores these data in the RGB field."                                                                                                             
#> [20] "LidarHistogram: Creates a histogram of LiDAR data."                                                                                                                                                                            
#> [21] "LidarIdwInterpolation: Interpolates LAS files using an inverse-distance weighted (IDW) scheme. When the input/output parameters are not specified, the tool interpolates all LAS files contained within the working directory."
#> [22] "LidarInfo: Prints information about a LiDAR (LAS) dataset, including header, point return frequency, and classification data and information about the variable length records (VLRs) and geokeys."                            
#> [23] "LidarJoin: Joins multiple LiDAR (LAS) files into a single LAS file."                                                                                                                                                           
#> [24] "LidarKappaIndex: Performs a kappa index of agreement (KIA) analysis on the classifications of two LAS files."                                                                                                                  
#> [25] "LidarNearestNeighbourGridding: Grids LAS files using nearest-neighbour scheme. When the input/output parameters are not specified, the tool grids all LAS files contained within the working directory."                       
#> [26] "LidarPointDensity: Calculates the spatial pattern of point density for a LiDAR data set. When the input/output parameters are not specified, the tool grids all LAS files contained within the working directory."             
#> [27] "LidarPointStats: Creates several rasters summarizing the distribution of LAS point data. When the input/output parameters are not specified, the tool works on all LAS files contained within the working directory."          
#> [28] "LidarRemoveDuplicates: Removes duplicate points from a LiDAR data set."                                                                                                                                                        
#> [29] "LidarRemoveOutliers: Removes outliers (high and low points) in a LiDAR point cloud."                                                                                                                                           
#> [30] "LidarSegmentation: Segments a LiDAR point cloud based on normal vectors."                                                                                                                                                      
#> [31] "LidarSegmentationBasedFilter: Identifies ground points within LiDAR point clouds using a segmentation based approach."                                                                                                         
#> [32] "LidarTINGridding: Creates a raster grid based on a Delaunay triangular irregular network (TIN) fitted to LiDAR points."                                                                                                        
#> [33] "LidarThin: Thins a LiDAR point cloud, reducing point density."                                                                                                                                                                 
#> [34] "LidarThinHighDensity: Thins points from high density areas within a LiDAR point cloud."                                                                                                                                        
#> [35] "LidarTile: Tiles a LiDAR LAS file into multiple LAS files."                                                                                                                                                                    
#> [36] "LidarTileFootprint: Creates a vector polygon of the convex hull of a LiDAR point cloud. When the input/output parameters are not specified, the tool works with all LAS files contained within the working directory."         
#> [37] "LidarTophatTransform: Performs a white top-hat transform on a Lidar dataset; as an estimate of height above ground, this is useful for modelling the vegetation canopy."                                                       
#> [38] "NormalVectors: Calculates normal vectors for points within a LAS file and stores these data (XYZ vector components) in the RGB field."                                                                                         
#> [39] "SelectTilesByPolygon: Copies LiDAR tiles overlapping with a polygon into an output directory."