Natural neighbour interpolation

Creates a raster grid based on Sibson's natural neighbour method.

wbt_natural_neighbour_interpolation(
  input,
  output,
  field = NULL,
  use_z = FALSE,
  cell_size = NULL,
  base = NULL,
  clip = TRUE,
  wd = NULL,
  verbose_mode = FALSE
)

Arguments

input
Input vector points file.

output
Output raster file.

field
Input field name in attribute table.

use_z
Use the 'z' dimension of the Shapefile's geometry instead of an attribute field?.

cell_size
Optionally specified cell size of output raster. Not used when base raster is specified.

base
Optionally specified input base raster file. Not used when a cell size is specified.

clip
Clip the data to the convex hull of the points?.

wd
Changes the working directory.

verbose_mode
Sets verbose mode. If verbose mode is False, tools will not print output messages.

Value

Returns the tool text outputs.